+++ src/test/java/racing/util/RandomNumberGeneratorTest.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.util;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
